Renormalize X after selecting the correct key frames.
X is normalized on the full X range to select the key frames.
The way the Y value is interpolated between key frames requires the X value to be between 0 (left key frame) and 1 (right key frame).
Therefore it must be renormalized from the full range to the range between key frames.